+++ spec/lib/polish_geeks/dev_tools/commands/rspec_spec.rb
@@ -11,30 +11,61 @@ RSpec.describe PolishGeeks::DevTools::Commands::Rspec do
           .with('bundle exec rspec spec')
       end
 
-      it 'executes the command' do
-        subject.execute
-      end
+      it { subject.execute }
     end
   end
 
   describe '#valid?' do
-    context 'when there are some failures' do
-      before do
-        subject.instance_variable_set(:@output, '2 failures')
-      end
+    context 'when there are failures' do
+      before { subject.instance_variable_set(:@output, '2 failures') }
 
-      it 'returns false' do
-        expect(subject.valid?).to eq false
-      end
+      it { expect(subject.valid?).to eq false }
     end
 
-    context 'when there are not any failures' do
-      before do
-        subject.instance_variable_set(:@output, '0 failures')
+    context 'when there are no failures' do
+      before { subject.instance_variable_set(:@output, '0 failures') }
+
+      it { expect(subject.valid?).to eq true }
+    end
+
+    context 'when there are no failures' do
+      let(:config) { double }
+
+      context 'and disallow pending false' do
+        before do
+          expect(PolishGeeks::DevTools::Config)
+            .to receive(:config) { config }
+          expect(config)
+            .to receive(:rspec_disallow_pending) { false }
+          subject.instance_variable_set(:@output, '0 failures, 2 pending')
+        end
+
+        it { expect(subject.valid?).to eq true }
       end
 
-      it 'returns true' do
-        expect(subject.valid?).to eq true
+      context 'and disallow pending true' do
+        before do
+          expect(PolishGeeks::DevTools::Config)
+            .to receive(:config) { config }
+          expect(config)
+            .to receive(:rspec_disallow_pending) { true }
+        end
+
+        context 'and there are pendings' do
+          before do
+            subject.instance_variable_set(:@output, '0 failures, 2 pending')
+          end
+
+          it { expect(subject.valid?).to eq false }
+        end
+
+        context 'and there are no pendings' do
+          before do
+            subject.instance_variable_set(:@output, '0 failures, 0 pending')
+          end
+
+          it { expect(subject.valid?).to eq true }
+        end
       end
     end
   end
@@ -43,14 +74,28 @@ RSpec.describe PolishGeeks::DevTools::Commands::Rspec do
     context 'when we run rspec' do
       let(:label) { '10 examples, 5 failures, 2 pending' }
 
-      before do
-        subject.instance_variable_set(:@output, label)
-      end
+      before { subject.instance_variable_set(:@output, label) }
 
-      it 'returns the label' do
-        expect(subject.label).to eq 'Rspec (10 ex, 5 fa, 2 pe)'
-      end
+      it { expect(subject.label).to eq 'Rspec (10 ex, 5 fa, 2 pe)' }
     end
+  end
+
+  describe '#examples_count' do
+    before { subject.instance_variable_set(:@output, '10 examples') }
+
+    it { expect(subject.send(:examples_count)).to eq(10) }
+  end
+
+  describe '#failures_count' do
+    before { subject.instance_variable_set(:@output, '10 failures') }
+
+    it { expect(subject.send(:failures_count)).to eq(10) }
+  end
+
+  describe '#pending_count' do
+    before { subject.instance_variable_set(:@output, '10 pending') }
+
+    it { expect(subject.send(:pending_count)).to eq(10) }
   end
 
   describe '.generator?' do