+++ lib/polish_geeks/dev_tools/commands/rspec.rb
@@ -21,16 +21,35 @@ module PolishGeeks
 
         # @return [Boolean] true if there were no Rspec failures, false otherwise
         def valid?
-          output.include?('0 failures')
+          failures_count.zero? && disallow_pending_valid?
         end
 
         # @return [String] default label for this command
         def label
-          examples = output.scan(EXAMPLES_REGEXP).flatten.first.to_i
-          failures = output.scan(FAILURES_REGEXP).flatten.first.to_i
-          pending = output.scan(PENDING_REGEXP).flatten.first.to_i
+          "Rspec (#{examples_count} ex, #{failures_count} fa, #{pending_count} pe)"
+        end
 
-          "Rspec (#{examples} ex, #{failures} fa, #{pending} pe)"
+        private
+
+        # @return [Integer] how many examples
+        def examples_count
+          output.scan(EXAMPLES_REGEXP).flatten.first.to_i
+        end
+
+        # @return [Integer] how many failures
+        def failures_count
+          output.scan(FAILURES_REGEXP).flatten.first.to_i
+        end
+
+        # @return [Integer] how many pending
+        def pending_count
+          output.scan(PENDING_REGEXP).flatten.first.to_i
+        end
+
+        # @return [Boolean] true if pending specs count is zero or we allow
+        # pending tests, false otherwise
+        def disallow_pending_valid?
+          PolishGeeks::DevTools.config.rspec_disallow_pending == true ? pending_count.zero? : true
         end
       end
     end