+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - #17: Bump rubycritic dependency to 2.3.0
 - Bump rspec dependency to 3.4.1
 - Bump activemodel dependency to 4.2.5
+- #30: Allow to pass config file to brakeman
 - #18: Add bundler-audit which checks for vulnerable versions of gems
 - #21: Replace readme validator with required files validators, which allows us to
   define what files you want to require in your project