+++ CHANGELOG.md
@@ -49,6 +49,15 @@ Now you need to change that to
 
 - #12: Add rubocop-rspec which checks add code style checking for your RSpec files
 
+- #33: Add an option to disallow pending specs
+
+If you want to allow pending specs then
+```ruby
+  PolishGeeks::DevTools.setup do |config|
+    config.rspec_disallow_pending = false
+  end
+```
+
 ## 1.2.1
 - Extracted all errors to PolishGeeks::DevTools::Errors namespace
 - Add support for validators